Flip right-hand control groups left/right.
On the right-hand control groups, display the
attenuverter knob and CV input jack on the right
instead of the left. This allows the control groups
to better fit inside the pentagons.

Tweak the pentagon corner coordinates to better fit
around the control groups.

Still need to adjust the text label positions
and the layout of the output jacks and connector lines.

+++ src/plugin.hpp
@@ -17,8 +17,8 @@ extern Model* modelTubeUnit;
 struct SapphireControlGroup     // represents the combination: knob + CV input + attenuverter
 {
     std::string name;
-    float yGrid;
-    float xGrid;
+    int yGrid;
+    int xGrid;
     int paramId;
     int attenId;
     int inputId;
